requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
urllib3==1.26.18
colorama==0.4.6
argparse==1.4.0
python-dotenv==1.0.0
tqdm==4.66.1
PyYAML==6.0.1
lxml==4.9.3
html5lib==1.1